LayoutTests/http/tests/contentextensions/block-everything-unless-domain-redirect-expected.txt

CONSOLE MESSAGE: Content blocker prevented frame displaying http://localhost/contentextensions/resources/should-not-load.html from loading a resource from http://localhost/contentextensions/resources/should-not-load.html
